// src/types/api.ts
export interface Epic {
  /** The Shortcut application url for the Epic. */
  app_url: string
  /** The Epic's description. */
  description: string
  /** True/false boolean that indicates whether the Epic is archived or not. */
  archived: boolean
  /** A true/false boolean indicating if the Epic has been started. */
  started: boolean
  /** A string description of this resource. */
  entity_type: string
  /** An array of Labels attached to the Epic. */
  labels: LabelSlim[]
  /** `Deprecated:` use `member_mention_ids`. */
  mention_ids: string[]
  /** An array of Member IDs that have been mentioned in the Epic description. */
  member_mention_ids: string[]
  /** An array containing Group IDs and Group-owned story counts for the Epic's associated groups. */
  associated_groups: EpicAssociatedGroup[]
  /** The IDs of Projects related to this Epic. */
  project_ids: number[]
  /**
   * The number of stories in this epic which are not associated with a project.
   * @format int64
   */
  stories_without_projects: number
  /**
   * A manual override for the time/date the Epic was completed.
   * @format date-time
   */
  completed_at_override?: string | null
  /**
   * The ID of the associated productboard integration.
   * @format uuid
   */
  productboard_plugin_id?: string | null
  /**
   * The time/date the Epic was started.
   * @format date-time
   */
  started_at?: string | null
  /**
   * The time/date the Epic was completed.
   * @format date-time
   */
  completed_at?: string | null
  /** An array of IDs for Objectives to which this epic is related. */
  objective_ids: number[]
  /** The name of the Epic. */
  name: string
  global_id: string
  /** A true/false boolean indicating if the Epic has been completed. */
  completed: boolean
  /** A nested array of threaded comments. */
  comments: ThreadedComment[]
  /** The URL of the associated productboard feature. */
  productboard_url?: string | null
  /**
   * The Epic's planned start date.
   * @format date-time
   */
  planned_start_date?: string | null
  /** `Deprecated` The workflow state that the Epic is in. */
  state: string
  /**
   * `Deprecated` The ID of the Objective this Epic is related to. Use `objective_ids`.
   * @format int64
   */
  milestone_id?: number | null
  /**
   * The ID of the Member that requested the epic.
   * @format uuid
   */
  requested_by_id: string
  /**
   * The ID of the Epic State.
   * @format int64
   */
  epic_state_id: number
  /** An array of Label ids attached to the Epic. */
  label_ids: number[]
  /**
   * A manual override for the time/date the Epic was started.
   * @format date-time
   */
  started_at_override?: string | null
  /**
   * `Deprecated` The ID of the group to associate with the epic. Use `group_ids`.
   * @format uuid
   */
  group_id?: string | null
  /**
   * The time/date the Epic was updated.
   * @format date-time
   */
  updated_at?: string | null
  /** An array of Group IDs that have been mentioned in the Epic description. */
  group_mention_ids: string[]
  /**
   * The ID of the associated productboard feature.
   * @format uuid
   */
  productboard_id?: string | null
  /** An array of UUIDs for any Members you want to add as Followers on this Epic. */
  follower_ids: string[]
  /** An array of UUIDS for Groups to which this Epic is related. */
  group_ids: string[]
  /** An array of UUIDs for any members you want to add as Owners on this new Epic. */
  owner_ids: string[]
  /** This field can be set to another unique ID. In the case that the Epic has been imported from another tool, the ID in the other tool can be indicated here. */
  external_id?: string | null
  /**
   * The unique ID of the Epic.
   * @format int64
   */
  id: number
  /**
   * The Epic's relative position in the Epic workflow state.
   * @format int64
   */
  position: number
  /** The name of the associated productboard feature. */
  productboard_name?: string | null
  /**
   * The Epic's deadline.
   * @format date-time
   */
  deadline?: string | null
  /** A group of calculated values for this Epic. */
  stats: EpicStats
  /**
   * The time/date the Epic was created.
   * @format date-time
   */
  created_at?: string | null
}

export interface EpicAssociatedGroup {
  /**
   * The Group ID of the associated group.
   * @format uuid
   */
  group_id: string
  /**
   * The number of stories this Group owns in the Epic.
   * @format int64
   */
  associated_stories_count?: number
}
// ...
export interface EpicSlim {
  /** The Shortcut application url for the Epic. */
  app_url: string
  /** The Epic's description. */
  description?: string
  /** True/false boolean that indicates whether the Epic is archived or not. */
  archived: boolean
  /** A true/false boolean indicating if the Epic has been started. */
  started: boolean
  /** A string description of this resource. */
  entity_type: string
  /** An array of Labels attached to the Epic. */
  labels: LabelSlim[]
  /** `Deprecated:` use `member_mention_ids`. */
  mention_ids: string[]
  /** An array of Member IDs that have been mentioned in the Epic description. */
  member_mention_ids: string[]
  /** An array containing Group IDs and Group-owned story counts for the Epic's associated groups. */
  associated_groups: EpicAssociatedGroup[]
  /** The IDs of Projects related to this Epic. */
  project_ids: number[]
  /**
   * The number of stories in this epic which are not associated with a project.
   * @format int64
   */
  stories_without_projects: number
  /**
   * A manual override for the time/date the Epic was completed.
   * @format date-time
   */
  completed_at_override?: string | null
  /**
   * The ID of the associated productboard integration.
   * @format uuid
   */
  productboard_plugin_id?: string | null
  /**
   * The time/date the Epic was started.
   * @format date-time
   */
  started_at?: string | null
  /**
   * The time/date the Epic was completed.
   * @format date-time
   */
  completed_at?: string | null
  /** An array of IDs for Objectives to which this epic is related. */
  objective_ids: number[]
  /** The name of the Epic. */
  name: string
  global_id: string
  /** A true/false boolean indicating if the Epic has been completed. */
  completed: boolean
  /** The URL of the associated productboard feature. */
  productboard_url?: string | null
  /**
   * The Epic's planned start date.
   * @format date-time
   */
  planned_start_date?: string | null
  /** `Deprecated` The workflow state that the Epic is in. */
  state: string
  /**
   * `Deprecated` The ID of the Objective this Epic is related to. Use `objective_ids`.
   * @format int64
   */
  milestone_id?: number | null
  /**
   * The ID of the Member that requested the epic.
   * @format uuid
   */
  requested_by_id: string
  /**
   * The ID of the Epic State.
   * @format int64
   */
  epic_state_id: number
  /** An array of Label ids attached to the Epic. */
  label_ids: number[]
  /**
   * A manual override for the time/date the Epic was started.
   * @format date-time
   */
  started_at_override?: string | null
  /**
   * `Deprecated` The ID of the group to associate with the epic. Use `group_ids`.
   * @format uuid
   */
  group_id?: string | null
  /**
   * The time/date the Epic was updated.
   * @format date-time
   */
  updated_at?: string | null
  /** An array of Group IDs that have been mentioned in the Epic description. */
  group_mention_ids: string[]
  /**
   * The ID of the associated productboard feature.
   * @format uuid
   */
  productboard_id?: string | null
  /** An array of UUIDs for any Members you want to add as Followers on this Epic. */
  follower_ids: string[]
  /** An array of UUIDS for Groups to which this Epic is related. */
  group_ids: string[]
  /** An array of UUIDs for any members you want to add as Owners on this new Epic. */
  owner_ids: string[]
  /** This field can be set to another unique ID. In the case that the Epic has been imported from another tool, the ID in the other tool can be indicated here. */
  external_id?: string | null
  /**
   * The unique ID of the Epic.
   * @format int64
   */
  id: number
  /**
   * The Epic's relative position in the Epic workflow state.
   * @format int64
   */
  position: number
  /** The name of the associated productboard feature. */
  productboard_name?: string | null
  /**
   * The Epic's deadline.
   * @format date-time
   */
  deadline?: string | null
  /** A group of calculated values for this Epic. */
  stats: EpicStats
  /**
   * The time/date the Epic was created.
   * @format date-time
   */
  created_at?: string | null
}
// ...
export interface LabelSlim {
  /** The Shortcut application url for the Label. */
  app_url: string
  /** The description of the Label. */
  description?: string | null
  /** A true/false boolean indicating if the Label has been archived. */
  archived: boolean
  /** A string description of this resource. */
  entity_type: string
  /**
   * The hex color to be displayed with the Label (for example, "#ff0000").
   * @format css-color
   * @minLength 1
   * @pattern ^#[a-fA-F0-9]{6}$
   */
  color?: string | null
  /** The name of the Label. */
  name: string
  global_id: string
  /**
   * The time/date that the Label was updated.
   * @format date-time
   */
  updated_at?: string | null
  /** This field can be set to another unique ID. In the case that the Label has been imported from another tool, the ID in the other tool can be indicated here. */
  external_id?: string | null
  /**
   * The unique ID of the Label.
   * @format int64
   */
  id: number
  /**
   * The time/date that the Label was created.
   * @format date-time
   */
  created_at?: string | null
}
// ...
export interface StorySearchResults {
  /**
   * The total number of matches for the search query. The first 1000 matches can be paged through via the API.
   * @format int64
   */
  total: number
  /** A list of search results. */
  data: StorySearchResult[]
  /** The URL path and query string for the next page of search results. */
  next?: string | null
}
// ...
export interface Workflow {
  /** A description of the workflow. */
  description: string
  /** A string description of this resource. */
  entity_type: string
  /** An array of IDs of projects within the Workflow. */
  project_ids: number[]
  /** A map of the states in this Workflow. */
  states: WorkflowState[]
  /** The name of the workflow. */
  name: string
  /**
   * The date the Workflow was updated.
   * @format date-time
   */
  updated_at: string
  /** Indicates if an owner is automatically assigned when an unowned story is started. */
  auto_assign_owner: boolean
  /**
   * The unique ID of the Workflow.
   * @format int64
   */
  id: number
  /**
   * The ID of the team the workflow belongs to.
   * @format int64
   */
  team_id: number
  /**
   * The date the Workflow was created.
   * @format date-time
   */
  created_at: string
  /**
   * The unique ID of the default state that new Stories are entered into.
   * @format int64
   */
  default_state_id: number
}
// ...
export interface FullRequestParams extends Omit<RequestInit, "body"> {
  /** set parameter to `true` for call `securityWorker` for this request */
  secure?: boolean
  /** request path */
  path: string
  /** content type of request body */
  type?: ContentType
  /** query params */
  query?: QueryParamsType
  /** format of response (i.e. response.json() -> format: "json") */
  format?: ResponseFormat
  /** request body */
  body?: unknown
  /** base url */
  baseUrl?: string
  /** request cancellation token */
  cancelToken?: CancelToken
}

export type RequestParams = Omit<FullRequestParams, "body" | "method" | "query" | "path">

export interface ApiConfig<SecurityDataType = unknown> {
  baseUrl?: string
  baseApiParams?: Omit<RequestParams, "baseUrl" | "cancelToken" | "signal">
  securityWorker?: (securityData: SecurityDataType | null) => Promise<RequestParams | void> | RequestParams | void
  customFetch?: typeof fetch
}